code/measure-plensor/artifact/requirements.txt
APScheduler==3.10.4
awscrt==0.19.1
awsiotsdk==1.19.0
boto3==1.35.42
botocore==1.35.42
cffi==1.17.1
jmespath==1.0.1
numpy==2.1.2
pillow==11.0.0
pycparser==2.22
pyserial==3.5b0
python-dateutil==2.9.0.post0
pytz==2024.2
RPi.GPIO==0.7.1
s3transfer==0.10.3
scipy==1.14.1
six==1.16.0
soundfile==0.12.1
tzlocal==5.2
urllib3==2.2.3
watchdog==5.0.3
